修改playerId为long类型

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/LogPlayer.java
@@ -35,7 +35,7 @@ public class LogPlayer implements Serializable {
 	 * 玩家id
 	 */
 	@Excel(name = "玩家id", width = 15)
-    private java.lang.Integer playerId;
+    private java.lang.Long playerId;
 	
 	/**
 	 * 服务器id
